+++ events/lib/applications.js
@@ -45,7 +45,6 @@ exports.setApplication = async (req, res) => {
     delete req.body.board_comment;
     delete req.body.status;
 
-
     let application = await Application.findOne({ where: { event_id: req.event.id, user_id: req.user.id } });
     if (application) {
         await application.update(req.body);
@@ -55,6 +54,7 @@ exports.setApplication = async (req, res) => {
         req.body.body_name = req.user.bodies.find((b) => b.id === req.body.body_id).name;
         req.body.user_id = req.user.id;
         req.body.event_id = req.event.id;
+        req.body.email = req.user.user.email;
 
         application = await Application.create(req.body);
     }