+++ events/migrations/20190909203402-add-emails-to-application.js
@@ -1,0 +1,10 @@
+module.exports = {
+    up: async (queryInterface, Sequelize) => {
+        await queryInterface.addColumn(
+            'applications',
+            'email',
+            { type: Sequelize.STRING, allowNull: false, defaultValue: 'Not set.' }
+        );
+    },
+    down: async (queryInterface) => queryInterface.removeColumn('applications', 'email')
+};